+++ src/pages/Diary/components/DiarySpot/DiarySpotHeader.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom';
 import { Button } from '~/components/common';
 import { DiaryHeader } from '~/pages/Diary/components/DiaryCommon';
 import { DiarySpotContext } from '~/pages/Diary/contexts/DiarySpotContent';
-import useInfo from '~/pages/Diary/hooks/useInfo';
+import useInfo from '~/pages/Diary/hooks/Diary/useInfo';
 
 const DiarySpotHeader = () => {
   const locate = useLocation();
