+++ src/pages/Diary/hooks/Diary/useInfo.ts
@@ -1,0 +1,12 @@
+import useDiary from '~/pages/Diary/hooks/Diary/useDiary';
+
+const useInfo = () => {
+  const { info, setInfo } = useDiary();
+
+  return {
+    info,
+    setInfo,
+  };
+};
+
+export default useInfo;
