+++ src/pages/Diary/hooks/Diary/useMarkers.ts
@@ -1,0 +1,12 @@
+import useDiary from '~/pages/Diary/hooks/Diary/useDiary';
+
+const useMarkers = () => {
+  const { markers, setMarkers } = useDiary();
+
+  return {
+    markers,
+    setMarkers,
+  };
+};
+
+export default useMarkers;
