+++ src/pages/Diary/hooks/DiaryMap/useDiaryMap.ts
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { DiaryMapContext } from '~/pages/Diary/contexts/DiaryMapContext';
+
+const useDiaryMap = () => {
+  const diaryMapContext = useContext(DiaryMapContext);
+
+  if (!diaryMapContext) throw new Error('Cannot find DiaryMapProvider');
+
+  return diaryMapContext;
+};
+
+export default useDiaryMap;
